feat(chat): refactor and fix send chat
